Replace Incorrect data types in API Spec (#66)
* change int32

* change uint

* fixed ippool types

* add minimum to request objects

* merge with master

* fix format

+++ v1/pkg/client/model_ip_pool_stats.go
@@ -14,9 +14,9 @@ package client
 // IpPoolStats struct for IpPoolStats
 type IpPoolStats struct {
 	// Total number of IPs (pool capacity)
-	Total float32 `json:"Total,omitempty"`
+	Total int32 `json:"Total,omitempty"`
 	// Number of non allocated IPs
-	Available float32 `json:"Available,omitempty"`
+	Available int32 `json:"Available,omitempty"`
 	// Number of allocated IPs
-	InUse float32 `json:"InUse,omitempty"`
+	InUse int32 `json:"InUse,omitempty"`
 }
